@@ -8,33 +8,34 @@
     "prebuild": "yarn",
     "build": "gatsby build",
     "build:serve": "yarn build && yarn serve",
-    "develop": "gatsby develop --open",
+    "dev": "gatsby develop --open",
     "lint": "tsc && eslint . --ext .ts,.tsx --fix",
-    "start": "yarn develop",
+    "start": "yarn dev",
     "serve": "gatsby serve --open",
     "test": "jest",
     "assemble": "yarn lint && yarn test && yarn build"
   },
   "dependencies": {
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-brands-svg-icons": "^5.13.1",
-    "@fortawesome/free-solid-svg-icons": "^5.13.1",
-    "@fortawesome/react-fontawesome": "^0.1.11",
+    "@chakra-ui/gatsby-plugin": "^2.0.4",
+    "@chakra-ui/react": "^1.8.8",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
     "antd": "^4.8.5",
-    "array-move": "^3.0.1",
-    "axios": "^0.21.0",
+    "array-move": "^4.0.0",
+    "axios": "^0.26.1",
+    "framer-motion": "^6.3.0",
     "gatsby-background-image": "^1.1.1",
+    "gbimage-bridge": "^0.2.1",
     "gsap": "^3.3.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-helmet": "^6.1.0",
-    "react-sortable-hoc": "^1.9.1"
+    "react-icons": "^4.3.1",
+    "react-sortable-hoc": "^2.0.0"
   },
   "devDependencies": {
     "@types/react-helmet": "^6.1.5",
-    "gatsby": "^2.23.1",
+    "gatsby": "^4.12.1",
     "gatsby-plugin-antd": "^2.2.0",
     "gatsby-plugin-less": "^4.2.0",
     "less": "^3.9.0"
