@@ -16,16 +16,16 @@
     "@mui/material": "^5.0.3",
     "@mui/styles": "^5.0.1",
     "@react-three/fiber": "^8.0.8",
-    "next": "^12.1.4",
+    "next": "^12.1.5",
     "next-images": "^1.8.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
     "three": "^0.139.2"
   },
   "devDependencies": {
-    "@types/node": "^16.10.3",
-    "@types/react": "^17.0.43",
-    "@types/react-dom": "^17.0.14",
+    "@types/node": "^17.0.24",
+    "@types/react": "^18.0.5",
+    "@types/react-dom": "^18.0.1",
     "typescript": "^4.6.3"
   }
 }
