@@ -8,66 +8,73 @@
     "prebuild": "yarn",
     "build": "gatsby build",
     "build:serve": "yarn build && yarn serve",
-    "develop": "gatsby develop --open -S",
+    "dev": "gatsby develop --open",
     "lint": "tsc && eslint . --ext .ts,.tsx --fix",
-    "start": "yarn develop",
+    "start": "yarn dev",
     "serve": "gatsby serve --open",
     "test": "jest",
     "assemble": "yarn lint && yarn test && yarn build"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-decorators": "^7.6.0",
-    "@types/d3-timer": "^2.0.0",
-    "@types/history": "^4.7.6",
+    "@babel/plugin-proposal-decorators": "^7.17.9",
+    "@types/d3-timer": "^3.0.0",
+    "@types/history": "^5.0.0",
     "@types/lodash": "^4.14.181",
-    "@types/node": "^16.10.3",
-    "@types/react": "^17.0.43",
+    "@types/node": "^17.0.24",
+    "@types/react": "^18.0.5",
     "@types/react-collapse": "^5.0.0",
-    "@types/react-dom": "^17.0.14",
+    "@types/react-dom": "^18.0.1",
     "@types/react-fontawesome": "^1.6.3",
-    "@types/react-motion": "^0.0.29",
-    "@types/react-test-renderer": "^17.0.0",
+    "@types/react-motion": "^0.0.32",
+    "@types/react-test-renderer": "^18.0.0",
     "babel-plugin-transform-typescript-metadata": "^0.3.0",
-    "babel-preset-gatsby": "^0.7.0",
-    "gatsby": "^2.23.6",
+    "babel-preset-gatsby": "^2.12.1",
+    "canvas": "^2.9.1",
+    "gatsby": "^4.12.1",
     "gatsby-plugin-alias-imports": "^1.0.5",
-    "gatsby-plugin-create-client-paths": "^2.1.23",
-    "gatsby-plugin-manifest": "^2.4.12",
-    "gatsby-source-filesystem": "^2.3.12",
+    "gatsby-plugin-manifest": "^4.12.1",
+    "gatsby-source-filesystem": "^4.12.1",
     "gatsby-theme-core": "*",
-    "gatsby-theme-material-ui": "^1.0.10"
+    "gatsby-theme-material-ui": "^5.2.0",
+    "husky": "^7.0.4"
   },
   "dependencies": {
-    "@emotion/react": "^11.0.0",
-    "@emotion/styled": "^11.0.0",
-    "@material-ui/core": "^4.10.1",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.44",
-    "@mui-treasury/layout": "^4.2.3",
-    "axios": "^0.21.0",
-    "core-js": "^3.6.4",
-    "d3-timer": "^2.0.0",
-    "easy-peasy": "^4.0.1",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui-treasury/layout": "^5.0.0",
+    "@mui/icons-material": "^5.6.1",
+    "@mui/lab": "^5.0.0-alpha.77",
+    "@mui/material": "^5.6.1",
+    "@mui/styles": "^5.6.1",
+    "axios": "^0.26.1",
+    "core-js": "^3.21.1",
+    "d3-timer": "^3.0.1",
+    "easy-peasy": "^5.0.4",
     "enzyme": "^3.9.0",
-    "husky": "^4.2.3",
-    "informed": "^3.8.1",
-    "konva": "^7.0.3",
+    "gatsby-plugin-image": "^2.12.1",
+    "graphql": "^16.3.0",
+    "husky": "^7.0.4",
+    "informed": "^4.9.3",
+    "konva": "^8.3.5",
     "lodash": "^4.17.15",
-    "mobx": "^6.0.4",
-    "mobx-react": "^7.0.5",
-    "mobx-react-form": "^2.0.8",
+    "mobx": "^6.5.0",
+    "mobx-react": "^7.3.0",
+    "mobx-react-form": "^3.2.0",
     "mobx-remotedev": "^0.3.6",
-    "react": "^17.0.1",
+    "prettier": "^2.0.5",
+    "pretty-quick": "^3.1.0",
+    "react": "^18.0.0",
     "react-card-flip": "^1.1.5",
     "react-collapse": "^5.1.1",
-    "react-device-detect": "^1.12.1",
-    "react-dom": "^17.0.1",
+    "react-device-detect": "^2.2.2",
+    "react-dom": "^18.0.0",
     "react-helmet": "^6.0.0",
     "react-hot-loader": "^4.12.21",
-    "react-konva": "^17.0.2-5",
+    "react-konva": "^18.0.0-0",
     "react-motion": "^0.5.2",
-    "react-sortable-hoc": "^1.10.1",
-    "react-test-renderer": "^17.0.1",
-    "use-image": "^1.0.3"
+    "react-sortable-hoc": "^2.0.0",
+    "react-test-renderer": "^18.0.0",
+    "typescript": "^4.1.2",
+    "use-image": "^1.0.12"
   }
 }
