@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Container, CircularProgress } from '@material-ui/core';
+import { Container, CircularProgress } from '@mui/material';
 
 const Loading = () => (
   <Container>
