@@ -1,12 +1,10 @@
 import * as React from 'react';
-import { Typography } from '@material-ui/core';
+import { Typography } from '@mui/material';
 
-const Head = () => {
-  return (
-    <Typography variant="h6" noWrap>
-      React Showcase
-    </Typography>
-  );
-};
+const Head = () => (
+  <Typography variant="h6" noWrap>
+    React Showcase
+  </Typography>
+);
 
 export default Head;
