@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { SvgIconProps } from '@material-ui/core/SvgIcon';
+import { SvgIconProps } from '@mui/material/SvgIcon';
 
-import { ListItem, ListItemIcon, ListItemText } from '@material-ui/core';
+import { ListItem, ListItemIcon, ListItemText } from '@mui/material';
 import { Link } from 'gatsby';
 
 interface ListItemLinkProps {
@@ -11,17 +11,15 @@ interface ListItemLinkProps {
   label: string;
   className?: string;
 }
-const ListItemLink: React.FC<ListItemLinkProps> = (props) => {
-  return (
-    <ListItem button component={Link} to={props.to} className={props.className}>
-      {props.icon && (
-        <ListItemIcon>
-          <props.icon />
-        </ListItemIcon>
-      )}
-      <ListItemText primary={props.label} />
-    </ListItem>
-  );
-};
+const ListItemLink: React.FC<ListItemLinkProps> = (props) => (
+  <ListItem button component={Link} to={props.to} className={props.className}>
+    {props.icon && (
+      <ListItemIcon>
+        <props.icon />
+      </ListItemIcon>
+    )}
+    <ListItemText primary={props.label} />
+  </ListItem>
+);
 
 export default ListItemLink;
