@@ -1,4 +1,4 @@
-import { Divider, List } from '@material-ui/core';
+import { Divider, List } from '@mui/material';
 import {
   Home,
   List as ListIcon,
@@ -7,10 +7,10 @@ import {
   Movie,
   Domain,
   Gavel,
-} from '@material-ui/icons';
+} from '@mui/icons-material';
 
 import React from 'react';
-import { css } from '@emotion/core';
+import { css } from '@emotion/css';
 import ListItemLink from './ListItemLink';
 import ListItemDropDown from './ListItemDropdown';
 
@@ -20,46 +20,44 @@ const styles = {
   `,
 };
 
-const SideBar = () => {
-  return (
-    <nav>
-      <List>
-        <ListItemLink to="/" icon={Home} label="Home" />
-      </List>
+const SideBar = () => (
+  <nav>
+    <List>
+      <ListItemLink to="/" icon={Home} label="Home" />
+    </List>
 
-      <Divider />
-      <List>
-        <ListItemLink to="/Todo" icon={ListIcon} label="Todo" />
+    <Divider />
+    <List>
+      <ListItemLink to="/Todo" icon={ListIcon} label="Todo" />
 
-        <ListItemLink to="/Shows" icon={Movie} label="Shows" />
+      <ListItemLink to="/Shows" icon={Movie} label="Shows" />
 
-        <ListItemLink to="/Domain" icon={Domain} label="Domain" />
+      <ListItemLink to="/Domain" icon={Domain} label="Domain" />
 
-        <ListItemLink to="/Cats" icon={Pets} label="Cats" />
+      <ListItemLink to="/Cats" icon={Pets} label="Cats" />
 
-        <ListItemDropDown icon={VideogameAsset} label="Games">
-          <ListItemLink
-            to="/Games/TicTacToe"
-            label="Tic-Tac-Toe"
-            css={styles.nestedStyle}
-          />
+      <ListItemDropDown icon={VideogameAsset} label="Games">
+        <ListItemLink
+          to="/Games/TicTacToe"
+          label="Tic-Tac-Toe"
+          className={styles.nestedStyle}
+        />
 
-          <ListItemLink
-            to="/Games/FlappyBird"
-            label="Flappy Bird"
-            css={styles.nestedStyle}
-          />
-          <ListItemLink
-            to="/Games/NumberGuess"
-            label="Number Guesser"
-            css={styles.nestedStyle}
-          />
-        </ListItemDropDown>
+        <ListItemLink
+          to="/Games/FlappyBird"
+          label="Flappy Bird"
+          className={styles.nestedStyle}
+        />
+        <ListItemLink
+          to="/Games/NumberGuess"
+          label="Number Guesser"
+          className={styles.nestedStyle}
+        />
+      </ListItemDropDown>
 
-        <ListItemLink icon={Gavel} to="/License" label="License" />
-      </List>
-    </nav>
-  );
-};
+      <ListItemLink icon={Gavel} to="/License" label="License" />
+    </List>
+  </nav>
+);
 
 export default SideBar;
