@@ -1,14 +1,11 @@
 import * as React from 'react';
 
 import SEO from 'src/components/modules/SEO';
-import { FormControl, TextField, Button, Divider } from '@material-ui/core';
+import { FormControl, TextField, Button, Divider } from '@mui/material';
 import { Form } from 'informed';
-import { css } from '@emotion/core';
+import { css } from '@emotion/css';
 
-interface State {
-  result: number;
-  number: number;
-}
+import { StaticImage } from 'gatsby-plugin-image';
 
 const catImage = css`
   width: 100%;
@@ -20,49 +17,44 @@ const catImage = css`
   }
 `;
 
-class Cats extends React.Component<{}, State> {
-  state = {
-    result: 200,
-    number: 200,
+const Cats: React.FC<any> = () => {
+  const [result, setResult] = React.useState(200);
+  const [number, setNumber] = React.useState(200);
+
+  const onChangeEvent = (e) => {
+    setResult(e.target.value);
+  };
+  const onSubmitEvent = () => {
+    setNumber(result);
   };
 
-  change = async (e) => {
-    this.setState({ result: e.target.value });
-  };
-
-  GetCat = () => {
-    this.setState((prevState) => ({ number: prevState.result }));
-  };
-
-  render() {
-    return (
-      <>
-        <SEO title="Cats" />
-        <div>
-          <h2>HTTP Cats</h2>
-          <img
-            src={`https://http.cat/${this.state.number}.jpg`}
-            alt="No cat found :("
-            css={catImage}
-          />
-          <Divider />
-          <Form onSubmit={this.GetCat}>
-            <FormControl>
-              <TextField
-                placeholder="Enter search term"
-                label="HTTP Status Code"
-                value={this.state.result}
-                onChange={this.change}
-              />
-              <Button variant="contained" color="primary">
-                Show me a cat
-              </Button>
-            </FormControl>
-          </Form>
-        </div>
-      </>
-    );
-  }
-}
+  return (
+    <>
+      <SEO title="Cats" />
+      <div>
+        <h2>HTTP Cats</h2>
+        <StaticImage
+          src={`https://http.cat/${number}.jpg`}
+          alt="No cat found :("
+          className={catImage}
+        />
+        <Divider />
+        <Form onSubmit={onSubmitEvent}>
+          <FormControl>
+            <TextField
+              placeholder="Enter search term"
+              label="HTTP Status Code"
+              value={result}
+              onChange={onChangeEvent}
+            />
+            <Button variant="contained" color="primary">
+              Show me a cat
+            </Button>
+          </FormControl>
+        </Form>
+      </div>
+    </>
+  );
+};
 
 export default Cats;
