@@ -1,16 +1,14 @@
 import React from 'react';
 import { Layout, Affix } from 'antd';
-import { css } from '@emotion/react';
+import { css } from '@emotion/css';
 import { Header } from './Header';
-import { Body } from './Body';
+import Body from './Body';
 import { Footer } from './Footer';
 import Banner from './Banner';
 
-const App = ({ children }) => <AppLayout>{children}</AppLayout>;
-
 const AppLayout = ({ children }) => (
   <Layout
-    css={css`
+    className={css`
       min-height: 100vh;
     `}
   >
@@ -21,7 +19,7 @@ const AppLayout = ({ children }) => (
       </Layout.Header>
     </Affix>
     <Layout.Content
-      css={css`
+      className={css`
         padding: 0 50px;
       `}
     >
@@ -32,5 +30,7 @@ const AppLayout = ({ children }) => (
     </Layout.Footer>
   </Layout>
 );
+
+const App = ({ children }) => <AppLayout>{children}</AppLayout>;
 
 export default App;
