@@ -1,10 +1,5 @@
 import React from 'react';
-import { Affix, Typography } from 'antd';
 import NavBar from './NavBar';
 
-export const Header = () => (
-  <>
-    <NavBar />
-  </>
-);
+export const Header = () => <NavBar />;
 export default Header;
