@@ -1,7 +1,7 @@
 import React from 'react';
 
 import SEO from 'src/components/SEO';
-import { css } from '@emotion/react';
+import { css } from '@emotion/css';
 
 import img from 'src/lib/img/cat-sunglasses.jpg';
 import fadeIn from 'src/components/animations/fadeIn';
@@ -11,16 +11,14 @@ const style = css`
   height: 100%;
   ${fadeIn}
 `;
-const IndexPage = () => (
-  <>
+const IndexPage = () => <>
     <SEO title="Home" />
     <h1>Cool Cats (WIP)</h1>
     <p>Pardon our appearance, our website is under construction </p>
 
     <div>
-      <img src={img} alt="Construction banner" css={style} />
+      <img src={img} alt="Construction banner" className={style} />
     </div>
   </>
-);
 
 export default IndexPage;
