@@ -1,1 +1,3 @@
-export { default } from './TextInput';
+import TextInput from './TextInput';
+
+export default TextInput;
