@@ -1,3 +1,3 @@
-export { default } from './TextInput';
+import TextInput from './TextInput';
 
-export { default as TextInput } from './TextInput';
+export default TextInput;
