@@ -1,5 +1,5 @@
 import React from 'react';
-import { css } from '@emotion/core';
+import { css } from '@emotion/css';
 import { FlipCard } from './FlipCard';
 
 const border = css`
@@ -13,27 +13,25 @@ const border = css`
   height: 130px;
 `;
 
-const CountdownUnit = (props: { time: number; label: string }) => {
-  return (
-    <div css={border}>
-      <h3
-        css={css`
-          margin: 0;
-        `}
-      >
-        {props.label}
-      </h3>
-      <div
-        css={css`
-          position: relative;
-          height: 100%;
-          overflow: hidden;
-        `}
-      >
-        <FlipCard time={props.time} />
-      </div>
+const CountdownUnit = (props: { time: number; label: string }) => (
+  <div className={border}>
+    <h3
+      className={css`
+        margin: 0;
+      `}
+    >
+      {props.label}
+    </h3>
+    <div
+      className={css`
+        position: relative;
+        height: 100%;
+        overflow: hidden;
+      `}
+    >
+      <FlipCard time={props.time} />
     </div>
-  );
-};
+  </div>
+);
 
 export default React.memo(CountdownUnit);
