@@ -1,15 +1,8 @@
 import * as React from 'react';
 
-class Charts extends React.Component {
-  state;
-
-  render() {
-    return (
-      <div>
-        <p>Charts</p>
-      </div>
-    );
-  }
-}
-
+const Charts = () => (
+  <div>
+    <p>Charts</p>
+  </div>
+);
 export default Charts;
