@@ -1,26 +1,24 @@
 import React from 'react';
 import SEO from 'src/components/modules/SEO';
 
-const License = () => {
-  return (
-    <>
-      <SEO title="License" />
-      <div>
-        <h1>License</h1>
-        <ul>
-          <li>
-            <a
-              href="http://www.transparentpng.com//details/phoenix-amazing-logo-image-download_21286.html"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Phoenix Amazing Logo Image Download @transparentpng.com
-            </a>
-          </li>
-        </ul>
-      </div>
-    </>
-  );
-};
+const License = () => (
+  <>
+    <SEO title="License" />
+    <div>
+      <h1>License</h1>
+      <ul>
+        <li>
+          <a
+            href="http://www.transparentpng.com//details/phoenix-amazing-logo-image-download_21286.html"
+            target="_blank"
+            rel="noopener noreferrer"
+          >
+            Phoenix Amazing Logo Image Download @transparentpng.com
+          </a>
+        </li>
+      </ul>
+    </div>
+  </>
+);
 
 export default License;
