@@ -2,8 +2,9 @@ import { Link } from 'gatsby';
 import showStore from 'src/state/stores/showStore';
 import React from 'react';
 import { Router } from '@reach/router';
-import { css } from '@emotion/core';
+import { css } from '@emotion/css';
 import { observer } from 'mobx-react-lite';
+
 // Image equalizer
 
 const imgStyle = css`
@@ -11,11 +12,12 @@ const imgStyle = css`
   height: 400px;
 `;
 
-const DetailsComponent = (props) => {
+const Details = (props) => {
+  console.log(props.params.id);
   const ShowStore = React.useContext(showStore);
 
   const show = ShowStore.shows.find(
-    (value) => value.id.toString() === props.id.toString()
+    (value) => value.id.toString() === props.params.id
   );
 
   return (
@@ -23,7 +25,7 @@ const DetailsComponent = (props) => {
       <img
         src={`https://image.tmdb.org/t/p/w500${show.poster_path}`}
         alt="where is the item"
-        css={imgStyle}
+        className={imgStyle}
       />
       <h1>{show.name || show.title}</h1>
       <p>{show.overview}</p>
@@ -31,11 +33,5 @@ const DetailsComponent = (props) => {
     </div>
   );
 };
-
-const Details = () => (
-  <Router>
-    <DetailsComponent path="Shows/Details/:id" />
-  </Router>
-);
 
 export default observer(Details);
