@@ -2,7 +2,7 @@ import { GatsbyConfig } from 'gatsby';
 
 const gatsbyConfig: GatsbyConfig = {
   // decrease bundle size
-  polyfill: false,
+  // polyfill: false,
   siteMetadata: {
     title: 'Placeholder Title',
     description: 'The placeholder in gatsby-theme-core',
@@ -11,11 +11,20 @@ const gatsbyConfig: GatsbyConfig = {
   plugins: [
     'gatsby-plugin-layout',
     'gatsby-plugin-eslint',
-    'gatsby-plugin-typescript',
     'gatsby-plugin-emotion',
     'gatsby-plugin-react-helmet',
+
+    // disabling image plugin for known issue at https://github.com/gatsbyjs/gatsby/issues/30516
+    //'gatsby-plugin-image',
+    {
+      resolve: 'gatsby-plugin-compile-es6-packages',
+      options: {
+        modules: ['gatsby-plugin-image'],
+      },
+    },
     'gatsby-transformer-sharp',
     'gatsby-plugin-sharp',
+    'gatsby-plugin-offline',
     {
       resolve: 'gatsby-plugin-manifest',
       options: {
@@ -28,7 +37,6 @@ const gatsbyConfig: GatsbyConfig = {
         icon: 'static/icon.png', // This path is relative to the root of the site.
       },
     },
-    'gatsby-plugin-offline',
     'gatsby-plugin-netlify',
     // add support for absolute paths for gatsby loader
     {
